269 use elf parameters instead of env vars (#270)
* get rid of ETH_TOKEN_MINT and COLLATERAL_POOL_BASE variables in source code

* edit run-proxy
